Removed some compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@2171 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/gui/track_sel.cpp
@@ -115,7 +115,7 @@ void TrackSel::updateScrollPosition()
 		end   = start+m_index_avail_tracks.size()-1;
 	}
 
-	for(unsigned int i=0; i<m_max_entries; i++)
+	for(unsigned int i=0; i<(unsigned int)m_max_entries; i++)
     {
 		if(i<start || i>end)
 		{
@@ -292,7 +292,7 @@ void TrackSel::displayImages(int selected_track)
 void TrackSel::update(float dt)
 {
     int indx = widget_manager->getSelectedWgt() - WTOK_TRACK0;
-    if(indx<0 || indx >= m_max_entries) 
+    if(indx<0 || indx >= (int)m_max_entries) 
     {
         widget_manager->update(dt);
         return;
